requirements.txt
pandas==2.2.2
PyPDF2==3.0.1
python_docx==1.1.2
python_Levenshtein==0.25.1
python_Levenshtein==0.25.1
scikit_learn==1.5.0
streamlit==1.35.0
torch==2.3.1
transformers==4.41.2
